feat(sdk/dotnet): fail-loud E2E, f64 fix, multi-target, NuGet publish
Maintainer follow-ups to bring the .NET SDK to sibling-SDK parity:

- fix(types): TokenInfo.IssuedAt and AuthEvent.Timestamp were typed `long`,
  but the server emits them as f64 (e.g. `1735689600.0`). System.Text.Json
  in strict mode refuses to bind a fractional number to an integer, so
  Cognito.GetTokensAsync/GetAuthEventsAsync threw against any real server
  holding a token or auth event. Retyped to `double` and locked with a
  FakeServerTests case using the exact `.0` wire form.

- test(e2e): replace the vacuous `if (_fc is null) return` skips with a
  self-spawning FakeCloudServer collection fixture that starts the binary,
  waits for health, and FAILS LOUD if the binary is missing or unreachable
  (matching the Java/Python/Go harnesses, not the weaker PHP skip). E2E can
  no longer pass green while asserting nothing. Honors FAKECLOUD_ENDPOINT to
  target an already-running server.

- feat(lifecycle): FakeCloudClient and HttpTransport now implement IDisposable
  so per-test clients release their HttpClient/connection pool.

- build: multi-target net8.0;net10.0 (net8 is the current LTS) instead of
  net10-only; add SourceLink, symbol packages, and deterministic CI builds;
  bump version 0.44.1 -> 0.44.3 to match the workspace and sibling SDKs.

- ci: the dotnet job no longer boots the server manually (the harness does);
  add the net8 SDK. Add a publish-nuget release job (dotnet pack + push,
  guarded by --skip-duplicate and the NUGET_API_KEY secret).

- docs: correct the ".NET 10+" requirement to ".NET 8 or later".

+++ sdks/dotnet/src/FakeCloud/Types.cs
@@ -500,7 +500,7 @@ public sealed record TokenInfo(
     string? Username,
     string? PoolId,
     string? ClientId,
-    long IssuedAt);
+    double IssuedAt);
 
 public sealed record TokensResponse(IReadOnlyList<TokenInfo>? Tokens);
 
@@ -513,7 +513,7 @@ public sealed record AuthEvent(
     string? Username,
     string? UserPoolId,
     string? ClientId,
-    long Timestamp,
+    double Timestamp,
     bool Success);
 
 public sealed record AuthEventsResponse(IReadOnlyList<AuthEvent>? Events);
